+++ src/client/logging/_global.ts
@@ -5,12 +5,12 @@
 import { isCI, isTestExecution } from '../common/constants';
 import { CallInfo } from '../common/utils/decorators';
 import { LogLevel } from './levels';
-import { configureLogger, createLogger, ILogger, LoggerConfig, logToAll } from './logger';
+import { configureLogger, createLogger, LoggerConfig, logToAll } from './logger';
 import { createTracingDecorator, LogInfo, TraceOptions, tracing as _tracing } from './trace';
 import { Arguments } from './util';
 
 const globalLogger = createLogger();
-initialize();
+initializeLogger();
 
 /**
  * Initialize the logger.
@@ -35,7 +35,7 @@ initialize();
  *   only if a file has been specified as an env variable.  Currently
  *   this is setup on CI servers.
  */
-function initialize() {
+export function initializeLogger() {
     const config: LoggerConfig = {};
     let nonConsole = false;
 
@@ -55,16 +55,7 @@ function initialize() {
         nonConsole = true;
     }
     configureLogger(globalLogger, config);
-
-    if (isCI && nonConsole) {
-        delete config.console;
-        // Send console.*() to the non-console loggers.
-        monkeypatchConsole(
-            // This is a separate logger that matches our config but
-            // does not do any console logging.
-            createLogger(config)
-        );
-    }
+    return config;
 }
 
 // Emit a log message derived from the args to all enabled transports.
@@ -116,42 +107,5 @@ export namespace traceDecorators {
         const opts = DEFAULT_OPTS;
         const level = LogLevel.Warn;
         return createTracingDecorator([globalLogger], { message, opts, level });
-    }
-}
-
-// Ensure that the console functions are bound before monkeypatching.
-import './transports';
-
-/**
- * What we're doing here is monkey patching the console.log so we can
- * send everything sent to console window into our logs.  This is only
- * required when we're directly writing to `console.log` or not using
- * our `winston logger`.  This is something we'd generally turn on, only
- * on CI so we can see everything logged to the console window
- * (via the logs).
- */
-function monkeypatchConsole(logger: ILogger) {
-    // The logging "streams" (methods) of the node console.
-    const streams = ['log', 'error', 'warn', 'info', 'debug', 'trace'];
-    const levels: { [key: string]: LogLevel } = {
-        error: LogLevel.Error,
-        warn: LogLevel.Warn
-    };
-    // tslint:disable-next-line:no-any
-    const consoleAny: any = console;
-    for (const stream of streams) {
-        // Using symbols guarantee the properties will be unique & prevents
-        // clashing with names other code/library may create or have created.
-        // We could use a closure but it's a bit trickier.
-        const sym = Symbol.for(stream);
-        consoleAny[sym] = consoleAny[stream];
-        // tslint:disable-next-line: no-function-expression
-        consoleAny[stream] = function () {
-            const args = Array.prototype.slice.call(arguments);
-            const fn = consoleAny[sym];
-            fn(...args);
-            const level = levels[stream] || LogLevel.Info;
-            logToAll([logger], level, args);
-        };
     }
 }
