+++ src/test/performanceTest.ts
@@ -24,6 +24,7 @@ import * as path from 'path';
 import * as request from 'request';
 import { EXTENSION_ROOT_DIR, PVSC_EXTENSION_ID } from '../client/common/constants';
 import { unzip } from './common';
+import { initializeLoggerForTests } from './test-logger';
 
 const NamedRegexp = require('named-js-regexp');
 const del = require('del');
@@ -37,6 +38,7 @@ enum Version {
     Release
 }
 
+initializeLoggerForTests();
 class TestRunner {
     public async start() {
         await del([path.join(tmpFolder, '**')]);
