+++ src/test/test-logger.ts
@@ -1,0 +1,58 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
+import { initializeLogger } from '../client/logging/_global';
+import { LogLevel } from '../client/logging/levels';
+import { createLogger, ILogger, logToAll } from '../client/logging/logger';
+
+export function initializeLoggerForTests() {
+    const config = initializeLogger();
+    if (process.env.VSC_PYTHON_LOG_FILE) {
+        delete config.console;
+        // Send console.*() to the non-console loggers.
+        monkeypatchConsole(
+            // This is a separate logger that matches our config but
+            // does not do any console logging.
+            createLogger(config)
+        );
+    }
+}
+
+// Ensure that the console functions are bound before monkeypatching.
+import '../client/logging/transports';
+
+/**
+ * What we're doing here is monkey patching the console.log so we can
+ * send everything sent to console window into our logs.  This is only
+ * required when we're directly writing to `console.log` or not using
+ * our `winston logger`.  This is something we'd generally turn on, only
+ * on CI so we can see everything logged to the console window
+ * (via the logs).
+ */
+function monkeypatchConsole(logger: ILogger) {
+    // The logging "streams" (methods) of the node console.
+    const streams = ['log', 'error', 'warn', 'info', 'debug', 'trace'];
+    const levels: { [key: string]: LogLevel } = {
+        error: LogLevel.Error,
+        warn: LogLevel.Warn
+    };
+    // tslint:disable-next-line:no-any
+    const consoleAny: any = console;
+    for (const stream of streams) {
+        // Using symbols guarantee the properties will be unique & prevents
+        // clashing with names other code/library may create or have created.
+        // We could use a closure but it's a bit trickier.
+        const sym = Symbol.for(stream);
+        consoleAny[sym] = consoleAny[stream];
+        // tslint:disable-next-line: no-function-expression
+        consoleAny[stream] = function () {
+            const args = Array.prototype.slice.call(arguments);
+            const fn = consoleAny[sym];
+            fn(...args);
+            const level = levels[stream] || LogLevel.Info;
+            logToAll([logger], level, args);
+        };
+    }
+}
